Struct puppetdb::client::Config

pub struct Config {
    pub server_urls: Vec<String>,
    pub cacert: String,
    pub cert: String,
    pub key: String,
}

Fields

server_urls

cacert

cert

key

Methods

impl Config

Struct to hold PuppetDB client configuration.

fn new(path: String, urls: String, cacert: String, cert: String, key: String) -> Config

Construct new client configuration, intended to be used with command flags where path is the path to the config to load if any of the other settings are empty.

fn load(path: String) -> Config

fn query(&self, query_str: String) -> Result

POSTs query_str (either AST or PQL) to configured PuppetDBs.